+++ src/Class/Rules/Pair.hpp
@@ -13,7 +13,7 @@ class Pair : public Rules{
 
         /* Getter */
         int getID() const;
-        int getScore() const;
+        float getScore() const;
         std::string getName() const;
 
         /* Setter */
